+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Adds a scraper for ESPN MMA data (UFC/PFL).
+- Matchup scraper for scraping event data (UFC/PFL).
+- Fight Details Scraper for scraping fight outcomes and stats (UFC/PFL).
 
 ## [0.15.1] - 2025-08-29
 ### Fixed